WebUI: make multi-value dialog more compact
Current representation of the multi-value dialog for the inventory was
taking a lot of vertical space, when we have plenty of horizontal
available.

This is the dialog that pops when a cell in the inventory has too much
data (because it is an aggregation of multiple values).

Reformatted to be a simple table with the keys and values side by side;
header in a single line with a close button represented by an X.