tcfl: fix not finding some interfaces
When running

 $ tcf healtcheck -i certs TARGETNAME

it was failing to find the certs interface because the information was
not downloaded; turns out the code in tcfl.ui_cli_targets was
optimizing too aggressively what to download.

This patch adds the projections_minize toggle to
tcfl.targets.run_fn_on_each_targetspec(), so we can tell it to just
download the whole inventory when we know it is needed...like for the
healthcheck.