ttbl.user_control: create secondary database directory if missing
Login operations were failing when secondary FSDB databases
directories did not exist. The code flow was incorrect and the
directory was not being created at all, always assuming it was an
inconsistency condition.

As so, login was failing with a 50x error when secondary FSDB storage
was enabled.

This patch ensures the FSDB directory for the user is created.