nreqs: add --generic-first to support installing in python venvs
When we were running nreqs.py in a venv environment, it was trying to
install distro general packages -- but then the virtual environment
could not reach them.

So add a switch to have nreqs.py install inside the virtual
environment by installing first with generic methods, such as pip; if
invoked with the right pythong ($VIRTUAL_ENV/bin/python nreqs.py
install ...), then it'll install in the virtual environment.