WebUI: add a UI widget to manage execution scripts in a remote pipeline against a target
Documentation described in doc/09-webui-widget-runner.rst--no good way
to break this more--read that for architectual implementation info and
deployment instructions.

This assumes you have the pipeline in place.

A modification in ttbd/ui/templates/target.html injects
ttbd/ui/templates/widget-target-runner.html, which will, from each
declaration in the inventory of a runner include an instantation of
ttbd/ui/templates/widget-target-runner-one.html.

That will drive code in ttbd/ui/static/js/widget-target-runner.js,
which is meant to interact with the remote pipelines (only jenkins
supported for now).

Changes in ui.py are to provide a full inventory accessible as a
javascript dictionary, which will allow removing a lot of it from the
hardcoded template we have now.