examples/qemu-bios: unify both testcases so it shows how N builds on N=1
This removes a lot of duplicated code and allows to build an example
of reuse.

We have the base test case which builds the new firmware version and
can provision one machine and test on one machine. We make it more
flexible to add a different string and look for it.

Then test_qemu_bios_N is re-made to inherit the simple one, reusing
the build instructions and then use that to firmware flash and test in
N machines at the same times.