tcfl.server_c.__discover_once: get better timing data on each server
Rework the flow of the thread pool so we cycle over the results as
soon as they are available; this allows us to get better timing data
on which servers are slower to respond to see if we need to take
action.

Change goes on using as_completed() which will get a result as long as
any sevrer completes instead of waiting for all to complete and then
getting the results.