WebUI/widget-runner:js_runner_jenkins_log_parse simplify state keeping and fix missing lines
The state was being kept in HTML's element, which was kind of messy.

Simplify by keeping the state in the stack itself and instead of doing
async calls, which will consume the stack, change to use setTimeout()
to keep stacking calls until the build is done.

Also fix how we parse out the lines from the GET call, since we were
over counting some lines and as a consequence dropping some lines
randomly.