WebUI/widget-runner: fix/simplify regexes to avoid recursion errors/locks
The way the regexes for TCF were too complex and too greedy because of
using * in the subcase string (where it is not needed, since we always
need a subcase) and ".*" is not really correct, but <not whitespace>*
(\S) to avoid . taking the rest of the string.

This reduces the amount of backtracking needed and removes the
recursion errors in Firefox and blocks in chrome.